requirements.txt
# Flask y extensiones
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0

# Base de datos
mysql-connector-python==8.0.32

# API de Gemini
google-generativeai==0.3.1

# Procesamiento de imágenes
Pillow==9.4.0
requests==2.28.2

# Utilidades
pydantic==2.1.1
python-dateutil==2.8.2

# Seguridad
PyJWT==2.6.0
bcrypt==4.0.1

# Desarrollo y Testing
pytest==7.3.1
black==23.3.0
flake8==6.0.0